Reorganize MultiKueue E2E tests
