+++ test/e2e/config/multikueue/baseline/kustomization.yaml
@@ -1,7 +1,7 @@
 apiVersion: kustomize.config.k8s.io/v1beta1
 kind: Kustomization
 resources:
-  - ../../../../dep-crds/training-operator/manifests/overlays/standalone
+  - ../../../../../dep-crds/training-operator/manifests/overlays/standalone
 patches:
   - path: patch_crds.yaml
     target:
