+++ test/e2e/config/multikueue/sequential/kustomization.yaml
@@ -2,7 +2,7 @@ apiVersion: kustomize.config.k8s.io/v1beta1
 kind: Kustomization
 
 resources:
-- ../../../../config/dev
+- ../../../../../config/dev
 
 replicas:
 - name: kueue-controller-manager
