+++ test/e2e/multikueue/sequential/e2e_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package multikueuesequential
+package sequential
 
 import (
 	"fmt"
